disabled button when submitting to prevent dos

+++ src/app/api/post/delete/route.ts
@@ -6,14 +6,22 @@ export async function POST(request: NextRequest) {
   const { postId } = await request.json();
 
   try {
-    const deletePost = await prisma.post.delete({
+    const post = await prisma.post.findFirst({
       where: {
         id: postId,
       },
     });
 
-    if (deletePost) {
-      return NextResponse.json({ success: true });
+    if (post) {
+      const deletePost = await prisma.post.delete({
+        where: {
+          id: post.id,
+        },
+      });
+
+      if (deletePost) {
+        return NextResponse.json({ success: true });
+      }
     }
   } catch (err) {
     return NextResponse.json({ success: false });
